+++ sdk-core/src/main/java/co/reachfive/identity/sdk/core/RedirectionActivity.kt
@@ -38,20 +38,28 @@ class RedirectionActivity : ComponentActivity() {
 
         const val RC_WEBLOGIN = 52557
 
+        const val RC_WEBLOGOUT = 52558
+
         fun isLoginRequestCode(code: Int): Boolean =
             setOf(RC_WEBLOGIN).contains(code)
+
+        fun isLogoutRequestCode(code: Int): Boolean =
+            setOf(RC_WEBLOGOUT).contains(code)
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        Log.d(TAG,"create")
 
         val urlString = intent.getStringExtra(URL_KEY)
         val codeVerifier = intent.getStringExtra(CODE_VERIFIER_KEY)
 
         val useWebView = intent.getBooleanExtra(USE_NATIVE_WEBVIEW, false)
         val originWebAuthn = intent.getStringExtra(ORIGIN_WEBAUTHN)
-        if (useWebView)
+
+        if (urlString == null) {
+            Log.d(TAG, "RedirectionActivity: no URL")
+            finish()
+        } else if (useWebView)
             launchWebView(codeVerifier, urlString, originWebAuthn)
         else
             launchCustomTab(urlString, codeVerifier)
@@ -105,7 +113,7 @@ class RedirectionActivity : ComponentActivity() {
 
     override fun onResume() {
         super.onResume()
-        Log.d(TAG,"onResume customTabStarted: $isCustomTabFlow")
+        Log.d(TAG, "RedirectionActivity onResume hasCustomTabStarted: $hasCustomTabStarted ; isCustomTabFlow: $isCustomTabFlow")
 
         // When Custom Tab returns, the Redirection Activity resumes and we need to end it.
         if (isCustomTabFlow && !hasCustomTabStarted) {
@@ -127,6 +135,8 @@ class RedirectionActivity : ComponentActivity() {
 
         val scheme = intent.getStringExtra(SCHEME)
         val url = newIntent.data
+        Log.d(TAG, "RedirectionActivity onNewIntent\nurl: $url\nscheme: $scheme")
+
         val isTargetR5 = intentClass == FQN && scheme != null && url.toString().startsWith(scheme)
         if (!isTargetR5) {
             Log.e(TAG, "Unrecognized intent class: $intentClass")
@@ -154,8 +164,11 @@ class RedirectionActivity : ComponentActivity() {
         super.onBackPressed()
 
         if (binding.webview.canGoBack()) {
+            Log.d(TAG, "RedirectionActivity onBackPressed go back")
             binding.webview.goBack()
         } else {
+            Log.d(TAG, "RedirectionActivity onBackPressed cancel")
+
             setResult(RESULT_CANCELED)
             finish()
         }
@@ -168,6 +181,7 @@ class RedirectionActivity : ComponentActivity() {
                 // regex : (reachfive://word character: [a-zA-Z_0-9]/callback)(any character zero or more times)
                 val pattern = Pattern.compile("^(reachfive:\\/\\/\\w+\\/callback)(.*)$")
                 val isTargetReachFive = pattern.matcher(url.toString()).matches()
+                Log.d(TAG, "WebViewClient isTargetReachFive: $isTargetReachFive")
 
                 if (isTargetReachFive) {
                     val intent = Intent()
@@ -179,7 +193,7 @@ class RedirectionActivity : ComponentActivity() {
                 } else return false
             }
 
-            Log.d(TAG, "WebView: unexpected empty url.")
+            Log.d(TAG, "WebViewClient: unexpected empty url.")
             finish()
             return true
         }